Fixed 2 bugs and improved information on tooltips and readme.
Fixed a bug that prevented the automatic save of the textures to folder expect the subfolder of the current location of the blend file.
Fixed a bug that prevented the selection of the separate levels when calculating the X axis from boundbox.
Improved couple tooltips for better visibility.
Add little more info in the readme.